RPI-22: [WIFI] Setting the security mode as WPAWPA2-PSK (TKIP/AES) is making  the WIFI client to connect to the SSID with the Authentication Mode as "WPA-Personal" instead of "WPA2-Personal".
Reason for change: Setting the security mode as WPAWPA2-PSK (TKIP/AES) is making
the WIFI client to connect to the SSID with the Authentication Mode as
"WPA-Personal" instead of "WPA2-Personal".

Test Procedure: Verified by setting the security mode as WPAWPA2-PSK (TKIP/AES) and
connecting the WIFI client to 2.4GHz band. Authentication Mode was "WPA2-Personal".

Risks: None

Change-Id: I94d4b85a9e7a0d787763bba99cc4a68bda973291
Signed-off-by: Avinash.Sharma@Cognizant.com